Starting out on a platform prototype
We now have two projects: "Hoops Prototype" and "Platforms Prototype".
Hoops is unchanged, and I've been working in Platforms.

Here's what I have so far:
- Platforms spawn on the staff with random notes and random durations.
The durations are multiples of beats, as defined in the BPM public
variable of game controller. The palette of notes used by the generator
can be adjusted in the "Allowed Notes" public array.