Fixed issue regarding 1st platform positioned incorrectly after change in speed
Did some extra testing and found that if the change in speed made a
platform become on screen and then off it destroyed it even if it was in
front of the player. Now such platforms must be passed (to the left) of
the player in order to be deleted. The null references after the game
closes are due to the order in which Unity deletes its objects; ignore
these?