Added range boxes to platforms
They're not color coded to show the upper and lower limits right now;
they're just translucent boxes encompassing the acceptable notes. Turns
out line renderers might not work that well for the gradient, since
doing so requires transposing the width and length of the line - and
when we do that, the box only shows up when the horizontal center is on
screen. So that's a drag.