Added graph of FFT transform
The level "PitchDebugger" draws a graph of what the pitch reader really
thinks we're doing. The x-axis represents frequency, the y-axis
intensity (as a fraction of the max intensity, rather than an absolute
unit; the strongest frequencywill reach all the way to the top of the
graph). Setting MinGraphedFreq and MaxGraphedFreq controls how much of
the spectrum it shows, while LabelSpacing determines how many
frequencies get labels on the x-axis.

I think this is really going to help! At a glance, it shows how
different vowels affect the harmonics.