Added second, parallaxed background layer
It looks like the best way to add a new layer is to create a new mesh,
move it slightly behind the previous one, and then give it a new
background (with the new image as the albedo component.

I currently have it set as cutout, which seems to work pretty well, but
requires changing the tint in order to lighten up the sprite.