Added neutral ticking sound for rests, adjusted sorting orders
FINALLY got the text meshes + Buddy's trail in the right place. @_@ That
was ridiculous. Turns out all renderers have a sortingOrder property,
even though SprtieRenderers are the only ones that expose it in the
editor. So from script, we can set the sorting order for line renderers,
text meshes, etc.