Fixed timing bug
Small errors were accumulating over the many coroutine iterations. This
led to the audio getting out of sync with the platforms/player. So,
here's the fix: Keep track of when each coroutine is SUPPOSED to fire.
Then, each time it does fire, take note of how off the timing is, and
take that into account when setting the next iteration's delay.
Note: the jump should always come before the audio cue, so I subtract
.15 seconds from its delay.