Added flats
Wow. That was... harder than I expected. So flats are PURELY COSMETIC,
do NOT try to add a flat to a song file! Any sharps have a 50/50 chance
of appearing as the corresponding flats. The names on the platforms are
updated, as are the positions.

Note: I thought it best to remove the sharp/flat part of the name from
the platform itself, since we don't actually have a flat symbol
character.