Added song ending, fixed glitch with y-pos loaded from song
We can add something nicer, but for the moment, the player just hits a
long rest at the end of songs, and runs off screen. Then it goes back to
the main menu.

As for the glitch, it's the same one from before, but my fix was being
overwritten in the AddNoteFromSong function.